refactor: useProduct 훅 네이밍 변경
